jserror

+++ frameworks/js/napi/source/location_napi_adapter.cpp
@@ -1666,6 +1666,15 @@ napi_value RemoveGnssGeofence(napi_env env, napi_callback_info info)
     NAPI_ASSERT(env, asyncContext != nullptr, "asyncContext is null.");
     asyncContext->fenceId_ = fenceId;
     asyncContext->callbackHost_ = FindCallbackInGnssGeofenceCallbackHostMap(fenceId);
+    if (asyncContext->callbackHost_ == nullptr) {
+        auto locationGnssGeofenceCallbackHost =
+            sptr<LocationGnssGeofenceCallbackNapi>(new LocationGnssGeofenceCallbackNapi());
+        napi_value object = nullptr;
+        JsObjToGeofenceTransitionCallback(env, object, locationGnssGeofenceCallbackHost);
+        auto callbackPtr = sptr<IGnssGeofenceCallback>(locationGnssGeofenceCallbackHost);
+        request->SetGeofenceTransitionCallback(callbackPtr->AsObject());
+        asyncContext->callbackHost_ = locationGnssGeofenceCallbackHost;
+    }
     if (napi_create_string_latin1(env, "removeGnssGeofence",
         NAPI_AUTO_LENGTH, &asyncContext->resourceName) != napi_ok) {
         GET_AND_THROW_LAST_ERROR(env);
@@ -1689,15 +1698,6 @@ void SetExecuteFuncForRemoveGnssGeofenceContext(GnssGeofenceAsyncContext* asyncC
         auto context = static_cast<GnssGeofenceAsyncContext*>(data);
         std::shared_ptr<GeofenceRequest> request = std::make_shared<GeofenceRequest>();
         request->SetFenceId(context->fenceId_);
-        if (context->callbackHost_ == nullptr) {
-            auto locationGnssGeofenceCallbackHost =
-                sptr<LocationGnssGeofenceCallbackNapi>(new LocationGnssGeofenceCallbackNapi());
-            napi_value object = nullptr;
-            JsObjToGeofenceTransitionCallback(context->env, object, locationGnssGeofenceCallbackHost);
-            auto callbackPtr = sptr<IGnssGeofenceCallback>(locationGnssGeofenceCallbackHost);
-            request->SetGeofenceTransitionCallback(callbackPtr->AsObject());
-            context->callbackHost_ = locationGnssGeofenceCallbackHost;
-        }
         context->errCode = g_geofenceClient->RemoveGnssGeofence(request);
         auto callbackHost = context->callbackHost_;
         if (callbackHost != nullptr) {
@@ -1726,7 +1726,7 @@ void SetCompleteFuncForRemoveGnssGeofenceContext(GnssGeofenceAsyncContext* async
             napi_get_undefined(context->env, &context->result[PARAM1]) == napi_ok) {
             LocationErrCode errCode = callbackHost->DealGeofenceOperationResult();
             if (errCode == ERRCODE_SUCCESS) {
-                RemoveCallbackToGnssGeofenceCallbackHostMap(context->fenceId_);
+                RemoveCallbackToGnssGeofenceCallbackHostMap(context->env, context->fenceId_);
             } else {
                 context->errCode = errCode;
             }
@@ -1984,13 +1984,15 @@ void AddCallbackToGnssGeofenceCallbackHostMap(int fenceId, sptr<LocationGnssGeof
     g_gnssGeofenceCallbackHostMap.insert(std::make_pair(fenceId, callbackHost));
 }
 
-void RemoveCallbackToGnssGeofenceCallbackHostMap(int fenceId)
+void RemoveCallbackToGnssGeofenceCallbackHostMap(napi_env env, int fenceId)
 {
     std::unique_lock<std::mutex> lock(g_gnssGeofenceCallbackHostMutex);
     auto iterForDelete = g_gnssGeofenceCallbackHostMap.find(fenceId);
     if (iterForDelete != g_gnssGeofenceCallbackHostMap.end()) {
-        if (iterForDelete->second != nullptr) {
+        if (iterForDelete->second != nullptr && env == iterForDelete->second->GetEnv()) {
             iterForDelete->second->DeleteHandler();
+        } else {
+            LBSLOGE(LOCATOR_STANDARD, "env is different when DeleteHandler");
         }
         g_gnssGeofenceCallbackHostMap.erase(iterForDelete);
     }
